refactor(connection): remove client property on connection model
