+++ src/views/window/Window.vue
@@ -41,9 +41,6 @@ export default class Window extends Vue {
     subscriptions: [],
     messages: [],
     unreadMessageCount: 0,
-    client: {
-      connected: false,
-    },
     ca: '',
     cert: '',
     key: '',
